✨ Separate into ZipExtractor, move applications folder check into UpdateInstaller
