+++ Loop/Localizable.xcstrings
@@ -3720,10 +3720,6 @@
         }
       }
     },
-    "Could not move the application to the Applications folder. Please do so manually for automatic updates to work." : {
-      "comment" : "Message shown in an alert when moving the application to the Applications folder fails.",
-      "isCommentAutoGenerated" : true
-    },
     "Credits" : {
       "comment" : "Section header shown in settings",
       "localizations" : {
@@ -5870,10 +5866,6 @@
         }
       }
     },
-    "Failed to Move Application" : {
-      "comment" : "Title of an alert that appears when moving the application to the Applications folder fails.",
-      "isCommentAutoGenerated" : true
-    },
     "Failed to resolve linked keybind" : {
       "comment" : "An error message displayed when a linked keybind cannot be resolved.",
       "localizations" : {
@@ -6785,10 +6777,6 @@
           }
         }
       }
-    },
-    "For automatic updates to work properly, this application should be in your Applications folder. Would you like to move it now?" : {
-      "comment" : "Informative text in an alert presented to the user when the default location for the app is not set. Asks the user if they would like to move the app to its default location.",
-      "isCommentAutoGenerated" : true
     },
     "Fourth Fourth" : {
       "comment" : "Window action",
@@ -10911,6 +10899,10 @@
         }
       }
     },
+    "Install to Applications" : {
+      "comment" : "Title of the first button in an alert that asks if the user wants to install an update to their Applications folder.",
+      "isCommentAutoGenerated" : true
+    },
     "Instant" : {
       "comment" : "Animation speed setting",
       "localizations" : {
@@ -11075,6 +11067,10 @@
           }
         }
       }
+    },
+    "Keep in Current Location" : {
+      "comment" : "Button title in an NSAlert that lets the user keep the update installed in its current location.",
+      "isCommentAutoGenerated" : true
     },
     "Keybinds" : {
       "comment" : "Section header shown in settings",
@@ -14447,6 +14443,10 @@
           }
         }
       }
+    },
+    "Loop is not in your Applications folder. Would you like to install the update to your Applications folder instead?" : {
+      "comment" : "Alert message presented to the user when the installed application is not in the Applications folder. Asks the user if they want to install the update there instead.",
+      "isCommentAutoGenerated" : true
     },
     "Loops left to unlock new icon" : {
       "localizations" : {
